disable button when creating a new inline comment
