+++ charmClient/hooks/comments.ts
@@ -1,0 +1,9 @@
+import type { ThreadCreate, ThreadWithComments, ThreadWithCommentsAndAuthors } from 'lib/threads/interfaces';
+
+import { usePOST } from './helpers';
+
+export function useCreateThread() {
+  return usePOST<Omit<ThreadCreate, 'userId'>, Promise<ThreadWithComments | ThreadWithCommentsAndAuthors>>(
+    '/api/threads'
+  );
+}
